change RepositoryId to int

+++ api/dtos/repositories.go
@@ -8,7 +8,7 @@ type RepositoryDTO struct {
 	OrganizationNodeId  string        `json:"organizationNodeId"`
 	Name                string        `json:"name"`
 	DisplayName         string        `json:"displayName"`
-	RepositoryId        int64         `json:"repositoryId"`
+	RepositoryId        int           `json:"repositoryId"`
 	Type                string        `json:"type"`
 	Description         string        `json:"description"`
 	URL                 string        `json:"url"`
